requirements.txt
numpy==1.22.3
oauthlib==3.2.0
opencensus==0.9.0
opencensus-context==0.1.2
opencensus-ext-azure==1.1.4
opencensus-ext-flask==0.8.0
opencv-python==4.6.0.66
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.4.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
passlib==1.7.4
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.1.0
plotly==5.7.0
pluggy==1.0.0
portalocker==2.4.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1656332401605/work
protobuf==4.21.2
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1653089181607/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1650904496387/work
PyJWT==2.4.0
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
pytest==7.1.2
python-barcode==0.14.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv==0.20.0
pytube==12.1.0
pytz==2022.1
PyYAML==6.0
pyzmq @ file:///home/builder/ci_310/pyzmq_1640794096902/work
qrcode==7.3.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn==0.0
SQLAlchemy==1.4.37
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1655315839047/work
tenacity==8.0.1
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.12.0
torchvision==0.13.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827254365/work
tqdm==4.64.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1655411388954/work
typing_extensions==4.2.0
urllib3==1.26.9
URLObject==2.4.3
utils==1.0.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
Werkzeug==2.1.1
WTForms==3.0.1
youtube-dl==2021.12.17
zipp==3.8.0